+++ CHANGELOG.md
@@ -34,6 +34,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Filter tables: results table is now fixed to the screen while scrolling facets, with sticky column headers. [#4271]
 - DwC Importer: When the Import/Retry button is pressed for a Checklist dataset, all rows are downloaded again instead of only the selected row.
 - Searching for Repositories now matches across multiple terms, e.g. 'dep hist' now matches '... Department ... History ...'
+- In filters, Taxon Name facets now default to 'Valid only' and 'Current only', and Taxon Determination facets now default to 'Current only'
 
 [#345]: https://github.com/SpeciesFileGroup/taxonpages/issues/345
 [#4271]: https://github.com/SpeciesFileGroup/taxonworks/issues/4271